Move `mountoptions` package to `github.com/awslabs/aws-s3-csi-driver/pkg/mountpoint`
